Created an inheritance hierarchy that makes more sense.
