+++ analyses/org.osate.analysis.resource.budgets/src/org/osate/analysis/resource/budgets/logic/AbstractResourceBudgetLogic.java
@@ -1,0 +1,192 @@
+/*
+ * <copyright>
+ * Copyright  2006 by Carnegie Mellon University, all rights reserved.
+ *
+ * Use of the Open Source AADL Tool Environment (OSATE) is subject to the terms of the license set forth
+ * at http://www.eclipse.org/legal/cpl-v10.html.
+ *
+ * NO WARRANTY
+ *
+ * ANY INFORMATION, MATERIALS, SERVICES, INTELLECTUAL PROPERTY OR OTHER PROPERTY OR RIGHTS GRANTED OR PROVIDED BY
+ * CARNEGIE MELLON UNIVERSITY PURSUANT TO THIS LICENSE (HEREINAFTER THE "DELIVERABLES") ARE ON AN "AS-IS" BASIS.
+ * CARNEGIE MELLON UNIVERSITY MAKES NO WARRANTIES OF ANY KIND, EITHER EXPRESS OR IMPLIED AS TO ANY MATTER INCLUDING,
+ * BUT NOT LIMITED TO, WARRANTY OF FITNESS FOR A PARTICULAR PURPOSE, MERCHANTABILITY, INFORMATIONAL CONTENT,
+ * NONINFRINGEMENT, OR ERROR-FREE OPERATION. CARNEGIE MELLON UNIVERSITY SHALL NOT BE LIABLE FOR INDIRECT, SPECIAL OR
+ * CONSEQUENTIAL DAMAGES, SUCH AS LOSS OF PROFITS OR INABILITY TO USE SAID INTELLECTUAL PROPERTY, UNDER THIS LICENSE,
+ * REGARDLESS OF WHETHER SUCH PARTY WAS AWARE OF THE POSSIBILITY OF SUCH DAMAGES. LICENSEE AGREES THAT IT WILL NOT
+ * MAKE ANY WARRANTY ON BEHALF OF CARNEGIE MELLON UNIVERSITY, EXPRESS OR IMPLIED, TO ANY PERSON CONCERNING THE
+ * APPLICATION OF OR THE RESULTS TO BE OBTAINED WITH THE DELIVERABLES UNDER THIS LICENSE.
+ *
+ * Licensee hereby agrees to defend, indemnify, and hold harmless Carnegie Mellon University, its trustees, officers,
+ * employees, and agents from all claims or demands made against them (and any related losses, expenses, or
+ * attorney's fees) arising out of, or relating to Licensee's and/or its sub licensees' negligent use or willful
+ * misuse of or negligent conduct or willful misconduct regarding the Software, facilities, or other rights or
+ * assistance granted by Carnegie Mellon University under this License, including, but not limited to, any claims of
+ * product liability, personal injury, death, damage to property, or violation of any laws or regulations.
+ *
+ * Carnegie Mellon University Software Engineering Institute authored documents are sponsored by the U.S. Department
+ * of Defense under Contract F19628-00-C-0003. Carnegie Mellon University retains copyrights in all material produced
+ * under this contract. The U.S. Government retains a non-exclusive, royalty-free license to publish or reproduce these
+ * documents, or allow others to do so, for U.S. Government purposes only pursuant to the copyright license
+ * under the contract clause at 252.227.7013.
+ * </copyright>
+ */
+package org.osate.analysis.resource.budgets.logic;
+
+import java.util.Iterator;
+
+import org.eclipse.emf.common.util.EList;
+import org.osate.aadl2.ComponentCategory;
+import org.osate.aadl2.NamedElement;
+import org.osate.aadl2.UnitLiteral;
+import org.osate.aadl2.instance.ComponentInstance;
+import org.osate.aadl2.instance.FeatureCategory;
+import org.osate.aadl2.instance.FeatureInstance;
+import org.osate.aadl2.instance.SystemInstance;
+import org.osate.aadl2.instance.SystemOperationMode;
+import org.osate.ui.handlers.AbstractAaxlHandler;
+import org.osate.xtext.aadl2.properties.util.GetProperties;
+
+abstract class AbstractResourceBudgetLogic extends AbstractLoggingLogic {
+	private final String prefixSymbol = "  ";
+	
+	protected int components = 0;
+	protected int budgetedComponents = 0;
+	
+	protected AbstractResourceBudgetLogic(AbstractAaxlHandler handler) {
+		super(handler);
+	}
+	
+	protected enum ResourceKind {
+		MIPS, RAM, ROM, Memory
+	}
+	
+	/**
+	 * calculate the budget of components with budgets, i.e., application
+	 * components and devices For application components they are required,
+	 * while for devices they are optional
+	 * 
+	 * @param ci component instance whose subtree is to be added up
+	 * @param rk Property Definition of property to be added
+	 * @param unit Unit in which the property value should be retrieved
+	 * @param somName String name of SOM (used in reporting)
+	 * @return double total, zero, if no budget, -1 if hardware only in
+	 *         substructure
+	 */
+	protected double sumBudgets(ComponentInstance ci, ResourceKind rk, UnitLiteral unit, final SystemOperationMode som,
+			String prefix) {
+		if (!ci.isActive(som)) {
+			return 0.0;
+		}
+		double subtotal = 0.0;
+		EList<ComponentInstance> subcis = ci.getComponentInstances();
+		boolean HWOnly = false;
+		boolean isSystemInstance = ci instanceof SystemInstance;
+		int subbudgetcount = 0;
+		int subcount = 0;
+		if (subcis.size() == 0) {
+			if (isHardware(ci)) {
+				return -1;
+			}
+		} else {
+			// track HWonly if subcomponents
+			HWOnly = true;
+		}
+		for (Iterator<ComponentInstance> it = subcis.iterator(); it.hasNext();) {
+			ComponentInstance subci = it.next();
+			double subresult = sumBudgets(subci, rk, unit, som, isSystemInstance ? "" : prefix + prefixSymbol);
+			if (subresult >= 0) {
+				HWOnly = false;
+				subtotal += subresult;
+				if (subci.getCategory() == ComponentCategory.DEVICE) {
+					if (subresult > 0) {
+						// only count device if it has a budget
+						subcount++;
+						subbudgetcount++;
+					}
+				} else {
+					// track how many non-devices and whether they have a budget
+					subcount++;
+					if (subresult > 0) {
+						subbudgetcount++;
+					}
+				}
+			}
+		}
+		if (HWOnly)
+			return -1;
+		double budget = getBudget(ci, rk);
+		if (rk.equals(ResourceKind.RAM) || rk.equals(ResourceKind.ROM) || rk.equals(ResourceKind.Memory)) {
+			double actualsize = getMemoryUseActual(ci, rk.name(), unit);
+			subtotal += actualsize;
+		}
+		String resourceName = ci.getCategory().getName();
+		String notes = "";
+		if (rk == ResourceKind.MIPS && ci.getCategory() == ComponentCategory.THREAD) {
+			subtotal = GetProperties.getThreadExecutioninMIPS(ci);
+		}
+		components = components + subcount;
+		budgetedComponents = budgetedComponents + subbudgetcount;
+		if (budget > 0 && subtotal > budget) {
+			notes = String.format("Error: subtotal/actual exceeds budget %.3f by %.3f " + unit.getName(), budget,
+					(subtotal - budget));
+		} else if (subtotal > 0 && subtotal < budget) {
+			notes = String.format(
+					resourceName + " " + ci.getInstanceObjectPath() + " total %.3f " + unit.getName()
+							+ " below budget %.3f " + unit.getName() + " (%.1f %% slack)",
+					subtotal, budget, (budget - subtotal) / budget * 100);
+		}
+		if (!isSystemInstance)
+			detailedLog(prefix, ci, budget, subtotal, resourceName, unit, notes);
+		return subtotal == 0 ? budget : subtotal;
+	}
+	
+	protected double getMemoryUseActual(ComponentInstance bci, String resourceName, UnitLiteral unit) {
+		double actualsize = 0.0;
+		if (resourceName.equals("ROM")) {
+			actualsize = GetProperties.getCodeSize(bci, unit);
+		} else if (resourceName.equals("RAM")) {
+			actualsize = GetProperties.getDataSize(bci, unit);
+			actualsize += GetProperties.getHeapSize(bci, unit);
+			actualsize += GetProperties.getStackSize(bci, unit);
+		} else {
+			actualsize = GetProperties.getDataSize(bci, unit);
+			actualsize += GetProperties.getHeapSize(bci, unit);
+			actualsize += GetProperties.getStackSize(bci, unit);
+			actualsize += GetProperties.getCodeSize(bci, unit);
+		}
+		return actualsize;
+	}
+	
+	private double getBudget(NamedElement ne, ResourceKind kind) {
+		switch (kind) {
+		case MIPS:
+			return GetProperties.getMIPSBudgetInMIPS(ne, 0.0);
+		case RAM:
+			return GetProperties.getRAMBudgetInKB(ne, 0.0);
+		case ROM:
+			return GetProperties.getROMBudgetInKB(ne, 0.0);
+		case Memory:
+			return GetProperties.getRAMBudgetInKB(ne, 0.0) + GetProperties.getROMBudgetInKB(ne, 0.0);
+		}
+		return 0.0;
+	}
+	
+	private boolean isHardware(ComponentInstance ci) {
+		ComponentCategory cat = ci.getCategory();
+		if (cat == ComponentCategory.BUS || cat == ComponentCategory.PROCESSOR
+				|| cat == ComponentCategory.VIRTUAL_PROCESSOR || cat == ComponentCategory.MEMORY)
+			return true;
+		if (cat == ComponentCategory.SYSTEM || cat == ComponentCategory.DEVICE) {
+			EList<FeatureInstance> el = ci.getFeatureInstances();
+			for (Iterator<FeatureInstance> it = el.iterator(); it.hasNext();) {
+				FeatureInstance fi = it.next();
+				if (fi.getCategory() != FeatureCategory.BUS_ACCESS) {
+					return false;
+				}
+			}
+			return true;
+		}
+		return false;
+	}
+}
